refactor(regions): apply review suggestions Co-authored-by: Paillat <me@paillat.dev>
